slightly better hires fix. (#495)
The idea is if the hr params match the prompt params, overwrite. But otherwise
force the hr prompt into a condition that won't blow up under combinatorial by
expanding the first or only hr prompt into all slots as needed.

Where this goes wrong is if an extension further up stack from us fills in
all_hr_prompts/all_hr_negative_prompts and we ignore the variations in other
slots. Not sure the best way to address that at this time.

Co-authored-by: Aarni Koskela <akx@iki.fi>